Show borrower in mybooks and minor refactoring
